pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.14",
    "aiohttp-jinja2>=1.6",
    "aiosqlite>=0.21.0",
    "imagehash>=4.3.2",
    "jinja2>=3.1.6",
    "pillow>=11.3.0",
    "psutil>=7.0.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "telegram>=0.0.1",
    "telethon>=1.40.0",
]
